feat: 新增可视化功能；新增未选和已选学生查看，支持随机分配; chore: 添加选课任务时，补充检查班级人数是否大于选课总人数的问题。

+++ backend/src/main/java/com/caixy/adminSystem/model/dto/courseSelectionInfo/CreateCourseSelectionRequest.java
@@ -51,7 +51,7 @@ public class CreateCourseSelectionRequest
     /**
      * 最小学分要求
      */
-    private BigDecimal minCredit;
+    private Double minCredit;
 
 
     @Data
